feat: tutorial remove generate proof and centralize proof submission in one state
