+++ web/assets/js/react/modules/Parity/AfterLevelCompletion.tsx
@@ -1,130 +1,14 @@
-import React, { useState } from "react";
+import React from "react";
 import { Button } from "../../components";
-import { generateCircomParityProof } from "./GenerateProof";
-import { VerificationData } from "../../types/aligned";
-import { Address } from "viem";
-import { SubmitProofModal } from "../../components/Modal/SubmitProof";
 import { ParityGameState } from "./types";
-
-type GameStatus = {
-	levelsBoards: number[][][];
-	userPositions: [number, number][][];
-};
 
 export const AfterLevelCompletion = ({
 	goToNextLevel,
-	levelBoards,
-	userPositions,
-	user_address,
-	payment_service_address,
-	batcher_url,
-	leaderboard_address,
-	currentGameConfig,
-	currentLevel,
 	setGameState,
 }: {
 	goToNextLevel: () => void;
-	levelBoards: number[][];
-	userPositions: [number, number][];
-	user_address: Address;
-	payment_service_address: Address;
-	batcher_url: string;
-	leaderboard_address: Address;
-	currentGameConfig: string;
-	currentLevel: number | null;
 	setGameState: (state: ParityGameState) => void;
 }) => {
-	const [proofVerificationData, setProofVerificationData] =
-		useState<VerificationData | null>(null);
-	const [open, setOpen] = useState(false);
-
-	const generateproofVerificationData = async () => {
-		const stored = localStorage.getItem("parity-game-data");
-
-		const gameData: { [key: string]: GameStatus } = stored
-			? JSON.parse(stored)
-			: {};
-
-		const currentLevelReached: GameStatus = gameData[currentGameConfig] || {
-			levelsBoards: [],
-			userPositions: [],
-		};
-
-		// If the current level is lower than the levels reached len, then replace the current and erase all the following levels
-		if (
-			currentLevel &&
-			currentLevelReached.levelsBoards.length > currentLevel
-		) {
-			currentLevelReached.levelsBoards =
-				currentLevelReached.levelsBoards.slice(0, currentLevel);
-			currentLevelReached.userPositions =
-				currentLevelReached.userPositions.slice(0, currentLevel);
-		}
-
-		// If the current level is equal to the levels reached length, then erase the current level data
-		if (
-			currentLevel &&
-			currentLevelReached.levelsBoards.length === currentLevel
-		) {
-			currentLevelReached.levelsBoards.pop();
-			currentLevelReached.userPositions.pop();
-		}
-
-		currentLevelReached.levelsBoards.push(levelBoards);
-		currentLevelReached.userPositions.push(userPositions);
-
-		// Save the data in local storage to use this level data for higher level proofs
-		gameData[currentGameConfig] = currentLevelReached;
-		localStorage.setItem("parity-game-data", JSON.stringify(gameData));
-
-		const submitproofVerificationData = await generateCircomParityProof({
-			user_address,
-			userPositions: currentLevelReached.userPositions,
-			levelsBoards: currentLevelReached.levelsBoards,
-		});
-
-		setProofVerificationData(submitproofVerificationData);
-		setOpen(true);
-	};
-
-	const saveLevelData = () => {
-		const stored = localStorage.getItem("parity-game-data");
-		const gameData: { [key: string]: GameStatus } = stored
-			? JSON.parse(stored)
-			: {};
-
-		const currentLevelReached: GameStatus = gameData[currentGameConfig] || {
-			levelsBoards: [],
-			userPositions: [],
-		};
-
-		// If the current level is lower than the levels reached len, then replace the current and erase all the following levels
-		if (
-			currentLevel &&
-			currentLevelReached.levelsBoards.length > currentLevel
-		) {
-			currentLevelReached.levelsBoards =
-				currentLevelReached.levelsBoards.slice(0, currentLevel);
-			currentLevelReached.userPositions =
-				currentLevelReached.userPositions.slice(0, currentLevel);
-		}
-
-		// If the current level is equal to the levels reached length, then erase the current level data
-		if (
-			currentLevel &&
-			currentLevelReached.levelsBoards.length === currentLevel
-		) {
-			currentLevelReached.levelsBoards.pop();
-			currentLevelReached.userPositions.pop();
-		}
-
-		currentLevelReached.levelsBoards.push(levelBoards);
-		currentLevelReached.userPositions.push(userPositions);
-
-		gameData[currentGameConfig] = currentLevelReached;
-		localStorage.setItem("parity-game-data", JSON.stringify(gameData));
-	};
-
 	return (
 		<div>
 			<div className="w-full h-full flex flex-col gap-4 items-center max-w-[500px]">
@@ -142,7 +26,6 @@ export const AfterLevelCompletion = ({
 							variant="arcade"
 							className="w-full"
 							onClick={() => {
-								saveLevelData();
 								goToNextLevel();
 							}}
 						>
@@ -152,7 +35,6 @@ export const AfterLevelCompletion = ({
 							variant="arcade"
 							className="w-full"
 							onClick={() => {
-								saveLevelData();
 								setGameState("home");
 							}}
 						>
@@ -161,16 +43,6 @@ export const AfterLevelCompletion = ({
 					</div>
 				</div>
 			</div>
-			<SubmitProofModal
-				modal={{ open, setOpen }}
-				batcher_url={batcher_url}
-				leaderboard_address={leaderboard_address}
-				payment_service_address={payment_service_address}
-				user_address={user_address}
-				userBeastSubmissions={[]}
-				proofToSubmitData={proofVerificationData}
-				gameName="parity"
-			/>
 		</div>
 	);
 };
