+++ web/assets/js/react/modules/Parity/GenerateProof.ts
